+++ modules/ClientRequest/Services/ClientRequestWorkflowService.php
@@ -8,12 +8,12 @@ use Illuminate\Database\UniqueConstraintViolationException;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
-use Modules\Shared\Process\Enums\ProcessStatus;
-use Modules\Shared\Process\Enums\ProcessStepStatus;
+use Modules\Process\Enums\ProcessStatus;
+use Modules\Process\Enums\ProcessStepStatus;
 use Modules\ClientRequest\Events\ClientRequestStatusChanged;
 use Modules\ClientRequest\Models\ClientRequest;
-use Modules\Shared\Process\Models\Process;
-use Modules\Shared\Process\Models\ProcessStep;
+use Modules\Process\Models\Process;
+use Modules\Process\Models\ProcessStep;
 use Modules\ProcedureSetting\Enums\ProcedureSettingType;
 use Modules\ProcedureSetting\Models\ProcedureSetting;
 use Modules\ProcedureSetting\Models\ProcedureSettingStep;
