+++ modules/Process/Commands/UpdateProcessCommand.php
@@ -1,0 +1,33 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Modules\Process\Commands;
+
+use Ramsey\Uuid\UuidInterface;
+
+class UpdateProcessCommand
+{
+    public function __construct(
+        private UuidInterface $id,
+        private string $name,
+    ) {
+    }
+
+    public function getId(): UuidInterface
+    {
+        return $this->id;
+    }
+
+    public function getName(): ?string
+    {
+        return $this->name;
+    }
+
+    public function toArray(): array
+    {
+        return [
+            'name' => $this->name,
+        ];
+    }
+}
