+++ modules/Process/Enums/ProcessStatus.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Modules\Process\Enums;
+enum ProcessStatus: string
+{
+    case Pending = 'pending';
+    case InProgress = 'in_progress';
+    case Completed = 'completed';
+    case Failed = 'failed';
+}
