+++ modules/Process/Requests/DeleteProcessRequest.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Modules\Process\Requests;
+
+use Illuminate\Foundation\Http\FormRequest;
+use Ramsey\Uuid\Uuid;
+
+class DeleteProcessRequest extends FormRequest
+{
+    public function rules(): array
+    {
+        return [];
+    }
+}
