+++ modules/Process/Requests/ExportProcessRequest.php
@@ -1,0 +1,28 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Modules\Process\Requests;
+
+use App\Http\Requests\BaseExportRequest;
+
+class ExportProcessRequest extends BaseExportRequest
+{
+    protected function getModelSpecificRules(): array
+    {
+        return [
+            'name' => 'sometimes|string|max:255',
+        ];
+    }
+
+    protected function getModelSpecificFilters(): array
+    {
+        $filters = [];
+        
+        if ($this->has('name')) {
+            $filters['name'] = $this->get('name');
+        }
+
+        return $filters;
+    }
+}
