+++ modules/Process/Requests/GetProcessListRequest.php
@@ -1,0 +1,19 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Modules\Process\Requests;
+
+use Illuminate\Foundation\Http\FormRequest;
+use Ramsey\Uuid\Uuid;
+
+class GetProcessListRequest extends FormRequest
+{
+    public function rules(): array
+    {
+        return [
+            'per_page' => 'integer',
+            'page' => 'integer',
+        ];
+    }
+}
