--- modules/Shared/Process/Enums/ProcessStepStatus.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Modules\Shared\Process\Enums;
-
-enum ProcessStepStatus: string
-{
-    case Pending = 'pending';
-    case Approved = 'approved';
-    case Rejected = 'rejected';
-}
