--- modules/Shared/Process/Filters/Shared/ProcessFilter.php
@@ -1,17 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Modules\Shared/Process\Filters;
-
-use BasePackage\Shared\Filters\SearchModelFilter;
-
-class Shared/ProcessFilter extends SearchModelFilter
-{
-       public $relations = [];
-
-        public function name($name)
-        {
-            return $this->where('name', $name);
-        }
-}
