--- modules/Shared/Process/Requests/ExportShared/ProcessRequest.php
@@ -1,28 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Modules\Shared/Process\Requests;
-
-use App\Http\Requests\BaseExportRequest;
-
-class ExportShared/ProcessRequest extends BaseExportRequest
-{
-    protected function getModelSpecificRules(): array
-    {
-        return [
-            'name' => 'sometimes|string|max:255',
-        ];
-    }
-
-    protected function getModelSpecificFilters(): array
-    {
-        $filters = [];
-        
-        if ($this->has('name')) {
-            $filters['name'] = $this->get('name');
-        }
-
-        return $filters;
-    }
-}
