--- modules/Shared/Process/Requests/GetShared/ProcessListRequest.php
@@ -1,19 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Modules\Shared/Process\Requests;
-
-use Illuminate\Foundation\Http\FormRequest;
-use Ramsey\Uuid\Uuid;
-
-class GetShared/ProcessListRequest extends FormRequest
-{
-    public function rules(): array
-    {
-        return [
-            'per_page' => 'integer',
-            'page' => 'integer',
-        ];
-    }
-}
