--- modules/Shared/Process/module.json
@@ -1,14 +1,0 @@
-{
-    "name": "Shared/Process",
-    "alias": "shared/process",
-    "description": "",
-    "keywords": [],
-    "priority": 0,
-    "providers": [
-        "Modules\\Shared/Process\\Providers\\Shared/ProcessServiceProvider"
-    ],
-    "aliases": {},
-    "files": [],
-    "requires": [],
-    "namespace": "Modules"
-}
